requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
requests
python-multipart
python-dotenv
azure-ai-inference==1.0.0b1
azure-core
python-dotenv
openai>=1.12.0
